refactor(react,state): fix import paths
