+++ packages/enty/package.json
@@ -22,7 +22,7 @@
     "watch": "yarn build -w"
   },
   "dependencies": {
-    "fronads": "^0.17.0",
+    "fronads": "^0.18.0",
     "unmutable": "^0.39.0"
   }
 }
